chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["David You <dyou3968@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
